+++ packages/std/src/query.rs
@@ -2,6 +2,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::encoding::Binary;
+use crate::types::{Coin, HumanAddr};
 
 pub type QueryResponse = Binary;
 
@@ -27,4 +28,27 @@ impl QueryResult {
             _ => false,
         }
     }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum QueryRequest {
+    // this queries the public API of another contract at a known address (with known ABI)
+    // msg is the json-encoded QueryMsg struct
+    // return value is whatever the contract returns (caller should know)
+    Contract {
+        contract_addr: HumanAddr,
+        msg: Binary, // we pass this in as Vec<u8> to the contract, so allow any binary encoding (later, limit to rawjson?)
+    },
+    // this calls into the native bank module
+    // return value is BalanceResponse
+    Balance {
+        address: HumanAddr,
+    },
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct BalanceResponse {
+    pub amount: Option<Vec<Coin>>,
 }