+++ packages/std/src/traits.rs
@@ -1,4 +1,7 @@
+use crate::api::{ApiError, ApiSystemError};
+use crate::encoding::Binary;
 use crate::errors::Result;
+use crate::query::QueryRequest;
 use crate::types::{CanonicalAddr, HumanAddr};
 
 #[cfg(feature = "iterator")]
@@ -34,18 +37,27 @@ pub trait Storage: ReadonlyStorage {
     fn remove(&mut self, key: &[u8]);
 }
 
-// Api are callbacks to system functions defined outside of the wasm modules.
-// This is a trait to allow Mocks in the test code.
-//
-// Currently it just supports address conversion, we could add eg. crypto functions here.
-// These should all be pure (stateless) functions. If you need state, you probably want
-// to use the Querier (TODO)
-//
-// We should consider if there is a way for modules to opt-in to only a subset of these
-// Api for backwards compatibility in systems that don't have them all.
+/// Api are callbacks to system functions defined outside of the wasm modules.
+/// This is a trait to allow Mocks in the test code.
+///
+/// Currently it just supports address conversion, we could add eg. crypto functions here.
+/// These should all be pure (stateless) functions. If you need state, you probably want
+/// to use the Querier (TODO)
+///
+/// We can use feature flags to opt-in to non-essential methods
+/// for backwards compatibility in systems that don't have them all.
 pub trait Api: Copy + Clone + Send {
     fn canonical_address(&self, human: &HumanAddr) -> Result<CanonicalAddr>;
     fn human_address(&self, canonical: &CanonicalAddr) -> Result<HumanAddr>;
+}
+
+pub trait Querier {
+    // Note: ApiError type can be serialized, and the below can be reconstituted over a WASM/FFI call.
+    // Since this is information that is returned from outside, we define it this way.
+    //
+    // ApiResult is a format that can capture this info in a serialized form. We parse it into
+    // a typical Result for the implementing object
+    fn query(&self, request: QueryRequest) -> Result<Result<Binary, ApiError>, ApiSystemError>;
 }
 
 // put them here to avoid so many feature flags